rozszerzony wybór regionów przez URL; poprawki headerów
